requirements.txt
line_bot_sdk==2.4.2
Flask==2.2.3
openai==0.27.2